MAINT|TST: Keep fixtures separate
