+++ src/apps/scdetect/test/fixture.cpp
@@ -1,4 +1,4 @@
-#include "utils.h"
+#include "fixture.h"
 
 #include <boost/program_options.hpp>
 #include <boost/test/unit_test.hpp>
